Changed project structure and namespaces
