+++ src/NuGetPackageManager.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28822.285
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NuGetPackageManager", "NuGetPackageManager\NuGetPackageManager.csproj", "{AE3DC056-6385-4DC6-83B5-3E40C7187FE0}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetPackageManager.Xaml", "NugetPackageManager.Xaml\NugetPackageManager.Xaml.csproj", "{190FB72F-899F-4AA2-A296-E2B812B97E0A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NugetPackageManager.Example", "NugetPackageManager.Example\NugetPackageManager.Example.csproj", "{29210134-578C-4414-A2C4-E6A301E029E5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NuGetPackageManager", "NugetPackageManager\NuGetPackageManager.csproj", "{190FB72F-899F-4AA2-A296-E2B812B97E0A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AE3DC056-6385-4DC6-83B5-3E40C7187FE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AE3DC056-6385-4DC6-83B5-3E40C7187FE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AE3DC056-6385-4DC6-83B5-3E40C7187FE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AE3DC056-6385-4DC6-83B5-3E40C7187FE0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{190FB72F-899F-4AA2-A296-E2B812B97E0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{190FB72F-899F-4AA2-A296-E2B812B97E0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{190FB72F-899F-4AA2-A296-E2B812B97E0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{190FB72F-899F-4AA2-A296-E2B812B97E0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{29210134-578C-4414-A2C4-E6A301E029E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29210134-578C-4414-A2C4-E6A301E029E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29210134-578C-4414-A2C4-E6A301E029E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29210134-578C-4414-A2C4-E6A301E029E5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
