+++ src/NuGetPackageManager/FodyWeavers.xml
@@ -1,5 +1,5 @@
 ﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <Catel />
-  <LoadAssembliesOnStartup />
   <ModuleInit />
+  <LoadAssembliesOnStartup />
+  <Catel />
 </Weavers>
