+++ src/NuGetPackageManager/ModuleInitializer.cs
@@ -1,5 +1,8 @@
-﻿using Catel.IoC;
-
+﻿using Catel.Configuration;
+using Catel.IoC;
+using NuGetPackageManager.Providers;
+using NuGetPackageManager.Services;
+using NuGetPackageManager.Model;
 
 /// <summary>
 /// Used by the ModuleInit. All code inside the Initialize method is ran as soon as the assembly is loaded.
@@ -13,5 +16,7 @@ public static class ModuleInitializer
     {
         var serviceLocator = ServiceLocator.Default;
 
+        serviceLocator.RegisterType<IConfigurationService, NugetConfigurationService>();
+        serviceLocator.RegisterType<IModelProvider<NugetFeed>, ModelProvider<NugetFeed>>();
     }
 }
