+++ src/NuGetPackageManager/NuGetPackageManager.csproj
@@ -1,4 +1,5 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras">
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
   <PropertyGroup>
     <TargetFrameworks>net462</TargetFrameworks>
     <AssemblyName>NuGetPackageManager</AssemblyName>
@@ -20,7 +21,20 @@
     <OutputType>WinExe</OutputType>
     <StartupObject />
     <NoWarn>$(NoWarn);SA1652</NoWarn>
+    <ApplicationIcon />
   </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="App.xaml" />
+    <None Remove="Resources\Add_8x_16x.png" />
+    <None Remove="Resources\cross-removebg.png" />
+    <None Remove="Resources\Download_grey_16x.png" />
+    <None Remove="Resources\Upload_gray_16x.png" />
+    <None Remove="Views\FeedDetailView.xaml" />
+    <None Remove="Views\MainWindow.xaml" />
+    <None Remove="Views\SettingsControlView.xaml" />
+    <None Remove="Views\SettingsWindow.xaml" />
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Catel.Fody" Version="4.0.0" />
@@ -30,20 +44,16 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="LoadAssembliesOnStartup.Fody" Version="4.0.0" />
+    <PackageReference Include="Microsoft-WindowsAPICodePack-Shell" Version="1.1.3.3" />
     <PackageReference Include="ModuleInit.Fody" Version="2.0.0" />
     <PackageReference Include="WpfAnalyzers" Version="2.2.0.2" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Update="NETStandard.Library" Version="2.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Properties\Resources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
+    <Resource Include="Resources\Add_8x_16x.png" />
+    <Resource Include="Resources\cross-removebg.png" />
+    <Resource Include="Resources\Download_grey_16x.png" />
+    <Resource Include="Resources\Upload_gray_16x.png" />
   </ItemGroup>
 
   <ItemGroup>
@@ -52,7 +62,5 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-  
-  <Import Project="$(MSBuildProjectDirectory)\..\Directory.build.shared.explicit.props" Condition="Exists('$(MSBuildProjectDirectory)\..\Directory.build.shared.explicit.props')" />
 
 </Project>
