+++ src/NuGetPackageManager/Views/MainWindow.xaml
@@ -1,17 +1,13 @@
 ﻿<catel:DataWindow x:Class="NuGetPackageManager.Views.MainWindow"
+        xmlns:catel="http://schemas.catelproject.com"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:NuGetPackageManager"
-        xmlns:catel="http://schemas.catelproject.com"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
     <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="9*"/>
-            <RowDefinition Height="1*"/>
-        </Grid.RowDefinitions>
-
+        <Button Width="180" Height="40"
+                Command="{Binding RunNuget}">Run Nuget Package Manager</Button>
     </Grid>
 </catel:DataWindow>
