+++ src/NuGetPackageManager/Views/MainWindow.xaml.cs
@@ -1,23 +1,27 @@
-﻿using Catel.Windows;
-using NuGetPackageManager.ViewModels;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
 
 namespace NuGetPackageManager.Views
 {
-    public partial class MainWindow : DataWindow
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Catel.Windows.DataWindow
     {
-        #region Constructors
-
         public MainWindow()
         {
             InitializeComponent();
         }
-
-        public MainWindow(MainViewModel viewModel)
-            : base(viewModel, DataWindowMode.Custom)
-        {
-            InitializeComponent();
-        }
-
-        #endregion
     }
 }
