--- src/NugetPackageManager.Example/NugetPackageManager.Example.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFrameworks>net462</TargetFrameworks>
-    <UseWPF>true</UseWPF>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Catel.MVVM" Version="5.10.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\NugetPackageManager.Xaml\NugetPackageManager.Xaml.csproj" />
-  </ItemGroup>
-
-</Project>
