--- src/NugetPackageManager.Xaml/FodyWeavers.xml
@@ -1,5 +1,0 @@
-﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <ModuleInit />
-  <LoadAssembliesOnStartup />
-  <Catel />
-</Weavers>
