--- src/NugetPackageManager.Xaml/NugetPackageManager.Xaml.csproj
@@ -1,44 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworks>net462</TargetFrameworks>
-    <UseWPF>true</UseWPF>
-    <ApplicationIcon />
-    <StartupObject />
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Resources\Add_8x_16x.png" />
-    <None Remove="Resources\cross-removebg.png" />
-    <None Remove="Resources\Download_grey_16x.png" />
-    <None Remove="Resources\Upload_gray_16x.png" />
-    <None Remove="Views\FeedDetailView.xaml" />
-    <None Remove="Views\SettingsControlView.xaml" />
-    <None Remove="Views\SettingsWindow.xaml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Catel.Fody" Version="4.0.0" />
-    <PackageReference Include="Catel.MVVM" Version="5.10.0" />
-    <PackageReference Include="Fody" Version="5.0.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="LoadAssembliesOnStartup.Fody" Version="4.0.0" />
-    <PackageReference Include="Microsoft-WindowsAPICodePack-Shell" Version="1.1.3.3" />
-    <PackageReference Include="ModuleInit.Fody" Version="2.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\NuGetPackageManager\NuGetPackageManager.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Resource Include="Resources\Add_8x_16x.png" />
-    <Resource Include="Resources\cross-removebg.png" />
-    <Resource Include="Resources\Download_grey_16x.png" />
-    <Resource Include="Resources\Upload_gray_16x.png" />
-  </ItemGroup>
-
-</Project>
