+++ src/NugetPackageManager/Constants.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NugetPackageManager.Xaml
+namespace NuGetPackageManager
 {
     public static class Constants
     {
