+++ src/NugetPackageManager/Resources/Add_8x_16x.png
