+++ src/NugetPackageManager/Resources/cross-removebg.png
