+++ src/NugetPackageManager/ViewModels/SettingsViewModel.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NugetPackageManager.Xaml.ViewModels
+namespace NuGetPackageManager.ViewModels
 {
     public class SettingsViewModel : ViewModelBase
     {
