+++ src/NugetPackageManager/Views/FeedDetailView.xaml
@@ -1,10 +1,10 @@
-﻿<catel:UserControl x:Class="NugetPackageManager.Xaml.Views.FeedDetailView"
+﻿<catel:UserControl x:Class="NuGetPackageManager.Views.FeedDetailView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:catel="http://schemas.catelproject.com"
-             xmlns:local="clr-namespace:NugetPackageManager.Xaml.Views"
+             xmlns:local="clr-namespace:NuGetPackageManager.Views"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
