+++ src/NugetPackageManager/Views/SettingsWindow.xaml
@@ -1,10 +1,10 @@
-﻿<catel:DataWindow x:Class="NugetPackageManager.Xaml.Views.SettingsWindow"
+﻿<catel:DataWindow x:Class="NuGetPackageManager.Views.SettingsWindow"
         xmlns:catel="http://schemas.catelproject.com"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:NugetPackageManager.Xaml.Views"
+        xmlns:local="clr-namespace:NuGetPackageManager.Views"
         mc:Ignorable="d"
         Title="SettingsWindow" MinHeight="400" MinWidth="500">
     <Grid>
