+++ src/NugetPackageManager/Views/SettingsWindow.xaml.cs
@@ -10,7 +10,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace NugetPackageManager.Xaml.Views
+namespace NuGetPackageManager.Views
 {
     /// <summary>
     /// Логика взаимодействия для SettingsWindow.xaml
